Flooring, Attempted Walls
Spent a few hours working on it and a little messing around :] just need
to make it.. clean and figure out why the textures assigned to the
sprites I push_back don't seem to be existing.

Perhaps the lifetime of the sprite is important, I figured it would just
make a copy, but what if the copy uses a reference to the texture of the
original? that's probably it.

I want to draw the floor separate so that I it can all be stored in a
fast VertexArray for every level, then I'll use a random level generator
to create the levels using destructible and/or invincible walls.