Walls
Walls have been added but they are kinda messy, I decided to get rid of
the 2D vector of sprites and just create them dynamically for now until
I have somewhere to actually implement them.

using 20x20 tiles is not a good idea because by generating a map where
every other tile is a wall you end up with one column and one row with
no/all wall, which makes it look ugly, it needs to be an odd number of
tiles in order to make a good map. I want to support really small
screens so the entire game has to fit in a resolution of 1024x768, at 21
tiles of 32 pixels that leaves 96 for the width which should be enough
for some kind of UI, otherwise there will be plenty of space at the top
to use even though the windows taskbar probably takes up ~32 pixels