Figured out issue with resource management
The MenuState contains the Menu's which contain the Buttons which contain the resources, the resoruce managers are being destroyd before the menu state is being destroyed, therefore the resource is attempting to destroy itself even though it no longer exists as it was cleaned up automatically when the manager was destroyed. There isn't really any fix, but I have made the buttons as shared_ptr's so they aren't copied all the time.